# rehketo-ui/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "McpServerCreate": {
        "properties": {
          "allowed_roles": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Roles",
            "type": "array"
          },
          "auth_token": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Token"
          },
          "auto_approve": {
            "default": false,
            "title": "Auto Approve",
            "type": "boolean"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "name": {
            "maxLength": 64,
            "pattern": "^[a-z0-9]+([_-][a-z0-9]+)*$",
            "title": "Name",
            "type": "string"
          },
          "url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "url",
          "allowed_roles"
        ],
        "title": "McpServerCreate",
        "type": "object"
      },
      "McpServerList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/McpServerOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "McpServerList",
        "type": "object"
      },
      "McpServerOut": {
        "properties": {
          "allowed_roles": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Roles",
            "type": "array"
          },
          "auto_approve": {
            "title": "Auto Approve",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "has_auth_token": {
            "title": "Has Auth Token",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "url",
          "has_auth_token",
          "allowed_roles",
          "enabled",
          "auto_approve",
          "created_at",
          "updated_at"
        ],
        "title": "McpServerOut",
        "type": "object"
      },
      "McpServerPatch": {
        "properties": {
          "allowed_roles": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Roles"
          },
          "auth_token": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Token"
          },
          "auto_approve": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Approve"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "title": "McpServerPatch",
        "type": "object"
      },
...
      "MySkillCreate": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "instructions": {
            "minLength": 1,
            "title": "Instructions",
            "type": "string"
          },
          "name": {
            "maxLength": 64,
            "pattern": "^[a-z0-9]+([_-][a-z0-9]+)*$",
            "title": "Name",
            "type": "string"
          },
          "trigger": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "name",
          "trigger",
          "instructions"
        ],
        "title": "MySkillCreate",
        "type": "object"
      },
      "MySkillList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MySkillOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "MySkillList",
        "type": "object"
      },
      "MySkillOut": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "editable": {
            "title": "Editable",
            "type": "boolean"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "display_name",
          "kind",
          "trigger",
          "instructions",
          "enabled",
          "source",
          "editable"
        ],
        "title": "MySkillOut",
        "type": "object"
      },
      "MySkillPatch": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "instructions": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "trigger": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger"
          }
        },
        "title": "MySkillPatch",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "description": "Session cookie set by /auth/login or /auth/devonly/login.",
        "in": "cookie",
        "name": "rehketo_session",
        "type": "apiKey"
      },
      "CSRFToken": {
        "description": "Double-submit CSRF token. Required on all unsafe methods except /auth/* and /healthz.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Rehketo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/mcp-servers": {
      "get": {
        "operationId": "list_servers_admin_mcp_servers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Servers",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "create_server_admin_mcp_servers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "CSRFToken": []
          }
        ],
        "summary": "Create Server",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/mcp-servers/{server_id}": {
      "delete": {
        "operationId": "delete_server_admin_mcp_servers__server_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Server Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "CSRFToken": []
          }
        ],
        "summary": "Delete Server",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "patch_server_admin_mcp_servers__server_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Server Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "CSRFToken": []
          }
        ],
        "summary": "Patch Server",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/me/skills": {
      "get": {
        "operationId": "list_my_skills_me_skills_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MySkillList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List My Skills",
        "tags": [
          "me"
        ]
      },
      "post": {
        "operationId": "create_my_skill_me_skills_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MySkillCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MySkillOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "CSRFToken": []
          }
        ],
        "summary": "Create My Skill",
        "tags": [
          "me"
        ]
      }
    },
    "/me/skills/{skill_id}": {
      "delete": {
        "operationId": "delete_my_skill_me_skills__skill_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "CSRFToken": []
          }
        ],
        "summary": "Delete My Skill",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "patch_my_skill_me_skills__skill_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MySkillPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MySkillOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "CSRFToken": []
          }
        ],
        "summary": "Patch My Skill",
        "tags": [
          "me"
        ]
      }
    },
...
  }
}